Stage 6

+++ src/main/java/gtPlusPlus/core/material/Material.java
@@ -582,7 +582,7 @@ public class Material {
                                         + aAverage
                                         + ".");
                         this.isRadioactive = true;
-                        this.vRadiationLevel = (byte) aAverage;
+                        this.vRadiationLevel = aAverage;
                     } else {
                         Logger.MATERIALS(this.getLocalizedName() + " is not radioactive.");
                         this.isRadioactive = false;
